Update1
int to float

+++ main.py
@@ -1,4 +1,4 @@
-triangle_side_a = int(input('Set triangle side size: '))
-triangle_height = int(input('Set triangle height length: '))
-triangle_square = int(triangle_side_a*triangle_height)/2
+triangle_side_a = float(input('Set triangle side size: '))
+triangle_height = float(input('Set triangle height length: '))
+triangle_square = float(triangle_side_a*triangle_height)/2
 print(triangle_square)
